Move datafeeds to tasks_io
