+++ src/backend/tasks_io/datafeeds/datafeed_base.py
@@ -10,7 +10,7 @@ from typing import (
 
 import requests
 
-from backend.common.datafeeds.parsers.parser_base import ParserBase
+from backend.tasks_io.datafeeds.parsers.parser_base import ParserBase
 
 
 TModel = TypeVar("TModel")
