+++ src/backend/tasks_io/datafeeds/parsers/fms_api/fms_api_event_alliances_parser.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, List, Optional
 
-from backend.common.datafeeds.parsers.json.parser_json import ParserJSON
 from backend.common.models.alliance import EventAlliance, EventAllianceBackup
+from backend.tasks_io.datafeeds.parsers.json.parser_json import ParserJSON
 
 
 class FMSAPIEventAlliancesParser(ParserJSON[List[EventAlliance]]):
