+++ src/backend/tasks_io/datafeeds/parsers/json/parser_paginated_json.py
@@ -1,7 +1,7 @@
 import abc
 from typing import Any, Dict, Optional, Tuple, TypeVar
 
-from backend.common.datafeeds.parsers.json.parser_json import ParserJSON
+from backend.tasks_io.datafeeds.parsers.json.parser_json import ParserJSON
 
 
 TParsedResponse = TypeVar("TParsedResponse")
