+++ src/backend/tasks_io/datafeeds/parsers/parser_html.py
@@ -3,7 +3,7 @@ import abc
 # import re
 from typing import TypeVar
 
-from backend.common.datafeeds.parsers.parser_base import ParserBase
+from backend.tasks_io.datafeeds.parsers.parser_base import ParserBase
 
 
 TParsedResponse = TypeVar("TParsedResponse")
